Change transform handler to use dependency injection
Instead of reading and writing the file directly, we inject the string from the already read file, and get back a string which we write to a file manually. This will make it easier to write a unit test for the transform method in the future, as we will be able to test for custom JSON strings and check the custom JSON outputs.