Add dummy method calls for error tests
If we don't do this then Visual Studio thinks that no test calls those methods and doesn't show next to that method that it has passing tests. So we add an explicit call from the test to make Visual Studio realise that we're testing that method, but because that method call throws an error, we cannot actually call it, therefore we put it in unreachable code and disable the corresponding warning.