Upgrade to System.CommandLine 2.0.0-beta6
This version of System.CommandLine has a completely different interface from the previous one, so it took a lot of work to make all the code compatible. That said, the new interface is MUCH better, so it was worthwhile. The flow is much more intuitive now, and we're not forced to use inheritance and understand extremely abstract concepts like command binders anymore. Now we just put mostly everything in static classes, which are the best kind of classes 🙂